Refactored home
