+++ lib/view/main_screen.dart
@@ -1,8 +1,8 @@
 import 'package:easy_study/presenter/hm_map.dart';
-import 'package:easy_study/presenter/home.dart';
 import 'package:easy_study/presenter/settings.dart';
 import 'package:easy_study/presenter/subject_add.dart';
 import 'package:easy_study/store/app_state.dart';
+import 'package:easy_study/view/home.dart';
 import 'package:easy_study/view/subject_overview.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
